+++ src/RadixDlt.NetworkGateway.PostgresIntegration/CommonDbContext.cs
@@ -115,6 +115,8 @@ internal abstract class CommonDbContext : DbContext
 
     public DbSet<NonFungibleIdStoreHistory> NonFungibleIdStoreHistory => Set<NonFungibleIdStoreHistory>();
 
+    public DbSet<NonFungibleIdLocationHistory> NonFungibleIdLocationHistory => Set<NonFungibleIdLocationHistory>();
+
     public DbSet<StateHistory> StateHistory => Set<StateHistory>();
 
     public DbSet<ValidatorPublicKeyHistory> ValidatorKeyHistory => Set<ValidatorPublicKeyHistory>();
@@ -188,7 +190,7 @@ internal abstract class CommonDbContext : DbContext
             .Properties<TokenAmount>()
             .HaveConversion<TokenAmountToBigIntegerConverter>()
             .HaveColumnType("numeric")
-            .HavePrecision(1000, 0);
+            .HavePrecision(1000);
 
         configurationBuilder
             .Properties<RadixAddress>()
@@ -414,6 +416,10 @@ internal abstract class CommonDbContext : DbContext
         modelBuilder
             .Entity<NonFungibleIdStoreHistory>()
             .HasIndex(e => new { e.NonFungibleResourceEntityId, e.FromStateVersion });
+
+        modelBuilder
+            .Entity<NonFungibleIdLocationHistory>()
+            .HasIndex(e => new { e.NonFungibleIdDataId, e.FromStateVersion });
 
         modelBuilder
             .Entity<StateHistory>()