+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Migrations/20230914131533_CreateNonFungibleIdLocationHistory.Designer.cs
@@ -81,8 +81,8 @@ using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 {
     [DbContext(typeof(MigrationsDbContext))]
-    [Migration("20230912144644_MarkSchemaDefiningColumnsAsRequired")]
-    partial class MarkSchemaDefiningColumnsAsRequired
+    [Migration("20230914131533_CreateNonFungibleIdLocationHistory")]
+    partial class CreateNonFungibleIdLocationHistory
     {
         /// <inheritdoc />
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
@@ -914,6 +914,38 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                     b.HasIndex("NonFungibleIdDataId", "FromStateVersion");
 
                     b.ToTable("non_fungible_id_data_history");
+                });
+
+            modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.NonFungibleIdLocationHistory", b =>
+                {
+                    b.Property<long>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("bigint")
+                        .HasColumnName("id");
+
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
+
+                    b.Property<long>("FromStateVersion")
+                        .HasColumnType("bigint")
+                        .HasColumnName("from_state_version");
+
+                    b.Property<bool>("IsDeleted")
+                        .HasColumnType("boolean")
+                        .HasColumnName("is_deleted");
+
+                    b.Property<long>("NonFungibleIdDataId")
+                        .HasColumnType("bigint")
+                        .HasColumnName("non_fungible_id_data_id");
+
+                    b.Property<long>("VaultEntityId")
+                        .HasColumnType("bigint")
+                        .HasColumnName("vault_entity_id");
+
+                    b.HasKey("Id");
+
+                    b.HasIndex("NonFungibleIdDataId", "FromStateVersion");
+
+                    b.ToTable("non_fungible_id_location_history");
                 });
 
             modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.NonFungibleIdStoreHistory", b =>
@@ -2112,6 +2144,16 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                                 .HasColumnType("sbor_type_kind[]")
                                 .HasColumnName("receipt_event_sbor_type_kinds");
 
+                            b1.Property<byte[][]>("EventSbors")
+                                .IsRequired()
+                                .HasColumnType("bytea[]")
+                                .HasColumnName("receipt_event_sbors");
+
+                            b1.Property<long[]>("EventSchemaEntityIds")
+                                .IsRequired()
+                                .HasColumnType("bigint[]")
+                                .HasColumnName("receipt_event_schema_entity_ids");
+
                             b1.Property<byte[][]>("EventSchemaHashes")
                                 .IsRequired()
                                 .HasColumnType("bytea[]")
@@ -2121,11 +2163,6 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                                 .IsRequired()
                                 .HasColumnType("bigint[]")
                                 .HasColumnName("receipt_event_type_indexes");
-
-                            b1.Property<byte[][]>("EventsSbors")
-                                .IsRequired()
-                                .HasColumnType("bytea[]")
-                                .HasColumnName("receipt_event_sbors");
 
                             b1.Property<string>("FeeDestination")
                                 .HasColumnType("jsonb")